requirements.txt
python-telegram-bot==20.6
pandas
opencv-python
PyMuPDF
Pillow
openpyxl
